backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
reportlab==4.0.9
Pillow==10.2.0
sendgrid==6.11.0
openai==1.12.0
anthropic==0.18.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
